+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/content/main.js
@@ -194,17 +194,14 @@ define([
 
             // getter for content data
             this.sandbox.on('sulu.content.contents.get-data', function(callback, reloadData, template) {
-                if (reloadData === true && template) {
+                if (this.options.id && reloadData === true && template) {
                     this.loadData(template).then(function() {
-                        // deep copy of object
                         callback(this.sandbox.util.deepCopy(this.data), this.preview);
-
                     }.bind(this));
 
                     return;
                 }
 
-                // deep copy of object
                 callback(this.sandbox.util.deepCopy(this.data), this.preview);
             }.bind(this));